services/benchmarks-service/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pandas==2.1.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0

